Move GlossaryTable into common
