+++ frontend/lint/module-boundaries.mjs
@@ -64,6 +64,7 @@ const elements = [
   createElement({ type: "shared", name: "collections", enforceOutgoing: true }),
   createElement({ type: "shared", name: "comments", enforceOutgoing: true }),
   createElement({ type: "shared", name: "data-grid", enforceOutgoing: true }),
+  createElement({ type: "shared", name: "data-studio", enforceOutgoing: true }),
   createElement({ type: "shared", name: "databases", enforceOutgoing: true }),
   createElement({ type: "shared", name: "history", enforceOutgoing: true }),
   createElement({ type: "shared", name: "hoc", enforceOutgoing: true }),
