+++ frontend/src/metabase/common/components/Glossary/Glossary.module.css
