+++ frontend/src/metabase/common/components/Glossary/GlossaryTable.tsx
